requirements.txt
pytest-playwright
asyncio
sqlalchemy==2.0.32
mysql-connector-python
schedule